redesigned the pr-slack-notification on manual merge to pr-manual-merge
